+++ packages/fabric/editor-bitbucket-transformer/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @atlaskit/editor-bitbucket-transformer
 
+## 2.1.0
+
+## 2.0.5
+
 ## 2.0.4
 
 ## 2.0.3