+++ packages/fabric/editor-cq/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/editor-cq
 
+## 6.11.2
+
+## 6.11.1
+
+## 6.11.0
+
+## 6.10.3
+
+## 6.10.2
+- [patch] added extension node [ec73cb8](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ec73cb8)
+- [patch] added extension node [ec73cb8](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ec73cb8)
+
 ## 6.10.1
 
 ## 6.10.0