+++ packages/fabric/editor-test-helpers/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @atlaskit/editor-test-helpers
 
+## 1.2.2
+- [patch] split extension node [4303d49](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4303d49)
+- [patch] split extension node [4303d49](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4303d49)
+
+## 1.2.1
+
+## 1.2.0
+
+## 1.1.3
+
+## 1.1.2
+- [patch] added extension node [ec73cb8](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ec73cb8)
+- [patch] added extension node [ec73cb8](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ec73cb8)
+
 ## 1.1.1
 
 ## 1.1.0