+++ packages/fabric/media-card/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/media-image
 
+## 22.4.0
+- [minor] make Card and CardView to work properly when percetanges are passed as dimensions [3178808](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/3178808)
+
+## 22.3.1
+
 ## 22.3.0
 - [minor] Update signature onClick event on filmstrip (renderer) [30bdfcc](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/30bdfcc)
 - [minor] Update signature onClick event on filmstrip (renderer) [dbced25](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/dbced25)